# Conflicts:
#	core/src/main/java/io/questdb/cairo/TableWriter.java
#	core/src/main/java/io/questdb/cairo/wal/SequencerImpl.java
#	core/src/main/java/io/questdb/cairo/wal/WalReader.java
#	core/src/main/java/io/questdb/cairo/wal/WalWriter.java
#	core/src/test/java/io/questdb/cairo/AbstractCairoTest.java
#	core/src/test/java/io/questdb/cairo/TableReaderTest.java

+++ core/src/main/java/io/questdb/cairo/wal/CopySegmentFileJob.java
@@ -46,7 +46,7 @@ public class CopySegmentFileJob {
             MemoryMA primaryColumn,
             MemoryMA secondaryColumn,
             @Transient Path walPath,
-            long newSegment,
+            int newSegment,
             CharSequence columnName,
             int columnType,
             long rowOffset,
